Lots of cleanup
